remove logging extensions
