+++ lib/structured_logger.rb
@@ -1,3 +1,14 @@
+# Ensure Tagged Logging formatter plays nicely with Ougai.
+# See also https://github.com/tilfin/ougai/wiki/Use-as-Rails-logger
+module ActiveSupport::TaggedLogging::Formatter
+  def call(severity, time, progname, data)
+    data = {msg: data.to_s} unless data.is_a?(Hash)
+    tags = current_tags
+    data[:tags] = tags if tags.present?
+    _call(severity, time, progname, data)
+  end
+end
+
 class StructuredLogger < Ougai::Logger
   include ActiveSupport::LoggerThreadSafeLevel
   include ActiveSupport::LoggerSilence
